240420 이효진 유효성 검사 완료

+++ src/main/java/com/example/demo/materials/MaterialInOutVO.java
@@ -13,7 +13,7 @@ public class MaterialInOutVO {
 	private String matLotCode;
 	private String matName;
 	private String companyName;
-	private Integer inoutCnt;
+	private Double inoutCnt;
 	private String inoutSep;
 	
 	// 검색 조건용
